+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1](https://github.com/LippertComponents/Orchestrator/compare/v1.3.0...v1.3.1) - 2019-06-26
+### Changed
+ - Add $modx as construct param to Deploy and DeployInterface
+ - Add check if custom deploy class implements DeployInterface
+
 ## [1.3.0](https://github.com/LippertComponents/Orchestrator/compare/v1.2.0...v1.3.0) - 2019-06-25
 
 ### Added